delete Http Context
