+++ srcs/Config/Config.hpp
@@ -10,16 +10,17 @@ class Config
     public:
         Config(const std::string& filepath);
 		HttpContext& getHttpContext();
-		static Config* getInstance();
 		const std::vector<std::string> getPorts();
 		~Config();
+        void addServerContext(const ServerContext& server);
+		void addDirective(const std::string& directive, const std::string& value,
+							const std::string& filepath, int line_number);
+		const std::map<std::string, std::vector<ServerContext> >& getServers() const;
+		const ServerContext& getServerContext(const std::string& port, const std::string& host) const;
 
     private:
-        HttpContext _http_context;
-        static Config* _instance;
-		int	redirectErrorLogFile(std::string errorLogFile);
-        int redirectAccessLogFile(std::string accessLogFile);
-        static const int NEW_FD;
+    	std::map<std::string, std::vector<ServerContext> > _servers;
+		std::map<std::string, std::string> _directives;      
 };
 
 #endif
