+++ srcs/Config/ConfigError.hpp
@@ -6,6 +6,7 @@
 
 enum ErrorType
 {
+	NEED_SERVER_CONTEXT,
 	DUPRICATE_DIRECTIVE,
 	NOT_ALLOWED_DIRECTIVE,
 	UNKOWN_DIRECTIVE,
